HUD sort of works

+++ SP1Framework/Object.h
@@ -15,7 +15,7 @@ public:
 	//set functions
 	void move(float, float);// change pos
 	void resize(int, int);// change size
-	void scale(int, int);// change size according to pivot point, multiplies the lengths
+	void scale(float, float);// change size according to pivot point, multiplies the lengths
 	void translate(float, float);// moves the object
 	void setPivot(int, int);
 	//get functions
